pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "livecodec"
version = "0.0.1"
description = "Lightweight and versatile neural codecs "
authors = [
    {name = "Dan Jacobellis", email = "danjacobellis@utexas.edu"}
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = [
    "torch",
    "einops",
    "PyWavelets",
    "timm",
    "torchvision",
    "pillow"
]
requires-python = ">=3.6"
